Feat:refactoring

+++ src/controllers/communityController.ts
@@ -1,6 +1,11 @@
 import { Request, Response } from "express";
 import { CommunityService } from "../services/communityService.js";
 import fs from "fs";
+import { asyncHandler } from "../middlewares/asyncHandler.js";
+import { STATUS_CODE } from "../utils/statusCode.js";
+import { buildResponse } from "../utils/builderResponse.js";
+import { AppError } from "../misc/AppError.js";
+import { commonErrors } from "../misc/commonErrors.js";
 
 interface MulterRequest extends Request {
   file: any;
@@ -9,73 +14,95 @@ interface MulterRequest extends Request {
 export class CommunityController {
   public communityService = new CommunityService();
 
-  public createPost = async (req: Request, res: Response) => {
-    try {
-      const { title, content, postType } = req.body;
+  public createPost = asyncHandler(async (req: Request, res: Response) => {
+    const user_id = req.id;
+    const { title, content, postType } = req.body;
+    if (req.file) {
+      const { originalname, path } = (req as MulterRequest).file;
+      const parts = originalname.split(".");
+      const ext = parts[parts.length - 1];
+      const newPath = path + "." + ext;
+      fs.renameSync(path, newPath);
 
-      if (req.file) {
-        const { originalname, path } = (req as MulterRequest).file;
-        const parts = originalname.split(".");
-        const ext = parts[parts.length - 1];
-        const newPath = path + "." + ext;
-        fs.renameSync(path, newPath);
-        const user_id: any = req.id;
-        const newPost = await this.communityService.createPost({
-          title,
-          content,
-          postType, //동행/함께
-          images: newPath,
-          user_id,
-        });
-        console.log("저장성공");
-        res.status(201).send(newPost);
-      } else {
-        const user_id: any = req.id;
-        const newPost = await this.communityService.createPost({
-          title,
-          content,
-          postType,
-          images: [],
-          user_id,
-        });
-        console.log("저장성공");
-        res.send(newPost);
-      }
-
-      console.log(req.body);
-    } catch (err) {
-      res.status(400).send(err);
+      const newPost = await this.communityService.createPost({
+        title,
+        content,
+        postType,
+        images: newPath,
+        user_id,
+      });
+      res.status(STATUS_CODE.OK).json(buildResponse(null, { newPost }));
+    } else {
+      const newPost = await this.communityService.createPost({
+        title,
+        content,
+        postType,
+        images: [],
+        user_id,
+      });
+      res.status(STATUS_CODE.OK).json(buildResponse(null, newPost));
     }
-  };
+  });
 
-  public getAllPosts = async (req: Request, res: Response) => {
-    try {
-      const posts = await this.communityService.findAllPost();
+  //멀터
+  // public createPost = asyncHandler(
+  //   async (req: Request, res: Response) => {
+  //     const { title, content, postType } = req.body;
 
-      res.send(posts);
-    } catch (err) {
-      res.status(400).send(err);
-    }
-  };
+  //     if (req.file) {
+  //       const { originalname, path } = (req as MulterRequest).file;
+  //       const parts = originalname.split(".");
+  //       const ext = parts[parts.length - 1];
+  //       const newPath = path + "." + ext;
+  //       fs.renameSync(path, newPath);
+  //       const user_id: any = req.id;
+  //       const newPost = await this.communityService.createPost({
+  //         title,
+  //         content,
+  //         postType,
+  //         images: newPath,
+  //         user_id,
+  //       });
+  //       console.log("저장성공");
+  //       res.status(STATUS_CODE.OK).json(buildResponse(null, newPost));
+  //     } else {
+  //       const user_id: any = req.id;
+  //       const newPost = await this.communityService.createPost({
+  //         title,
+  //         content,
+  //         postType,
+  //         images: [],
+  //         user_id,
+  //       });
+  //       console.log("저장성공");
+  //       res.status(STATUS_CODE.OK).json(buildResponse(null, newPost));
+  //     }
+  //   }
+  // );
 
-  public searchPost = async (req: Request, res: Response) => {
+  // 모든 게시물 조회
+  public getAllPosts = asyncHandler(async (req: Request, res: Response) => {
+    const posts = await this.communityService.findAllPost();
+    res.status(STATUS_CODE.OK).json(buildResponse(null, posts));
+  });
+  ////
+
+  //keyword 로 게시물 조회
+  public searchPost = asyncHandler(async (req: Request, res: Response) => {
     const { keyword } = req.query;
-    try {
-      const posts = await this.communityService.searchPost(keyword as string);
-      res.send(posts);
-    } catch (err) {
-      res.status(400).send(err);
-    }
-  };
 
-  public getPost = async (req: Request, res: Response) => {
+    const posts = await this.communityService.searchPost(keyword as string);
+    res.status(STATUS_CODE.OK).json(buildResponse(null, posts));
+  });
+
+  public getPost = asyncHandler(async (req: Request, res: Response) => {
     const { id } = req.params;
 
     const comment = await this.communityService.findByPostIdComment(id);
-
     const post = await this.communityService.indByPostIdPost(id);
-    res.send({ post, comment });
-  };
+
+    res.status(STATUS_CODE.OK).json(buildResponse(null, { post, comment }));
+  });
 
   public patchPost = async (req: Request, res: Response) => {
     const { id } = req.params;
@@ -109,34 +136,51 @@ export class CommunityController {
       res.status(400).send({ message: "오류 발생" });
     }
   };
-  public getPostByCategory = async (req: Request, res: Response) => {
-    const { category } = req.params;
-    try {
-      const categoryPost = await this.communityService.getPostByCat(category);
-      res.send(categoryPost);
-    } catch {}
-  };
+  //카테고리
+  public getPostByCategory = asyncHandler(
+    async (req: Request, res: Response) => {
+      const { category } = req.params;
+      if (!category) {
+        throw new AppError(
+          commonErrors.argumentError,
+          STATUS_CODE.BAD_REQUEST,
+          "BAD_REQUEST"
+        );
+      }
+      const categoryPost = await this.communityService.getPostByCat(
+        category as string
+      );
+      res.status(STATUS_CODE.OK).json(buildResponse(null, categoryPost));
+    }
+  );
 
-  public deletePost = async (req: Request, res: Response) => {
+  //게시물 삭제
+  public deletePost = asyncHandler(async (req: Request, res: Response) => {
     const { id } = req.params;
-    try {
-      // await PostModel.deleteOne({ _id: id });
-      await this.communityService.delete(id);
-      res.send({
-        message: "삭제가 완료되었습니다.",
-      });
-    } catch {
-      res.status(400).send({ message: "오류 발생" });
+    if (!id) {
+      throw new AppError(
+        commonErrors.argumentError,
+        STATUS_CODE.BAD_REQUEST,
+        "BAD_REQUEST"
+      );
     }
-  };
+    await this.communityService.delete(id);
+    res.status(STATUS_CODE.OK).json(buildResponse(null, null));
+  });
+  /////////
 
-  public getUserPosts = async (req: Request, res: Response) => {
+  // 유저 정보 가져오기
+  public getUserPosts = asyncHandler(async (req: Request, res: Response) => {
     const user_id: any = req.id;
-    try {
-      const userPosts = await this.communityService.getUserPosts(user_id);
-      res.status(200).send(userPosts);
-    } catch {
-      res.status(404).send({ message: "오류 발생" });
+    if (!user_id) {
+      throw new AppError(
+        commonErrors.requestValidationError,
+        STATUS_CODE.BAD_REQUEST,
+        "BAD_REQUEST"
+      );
     }
-  };
+
+    const userPosts = await this.communityService.getUserPosts(user_id);
+    res.status(STATUS_CODE.OK).json(buildResponse(null, userPosts));
+  });
 }
